// lib/enums.ts
export enum ListType {
  OPEN = 'open',
  WOMEN = 'women',
  JUNIORS = 'juniors',
  GIRLS = 'girls',
  MEN_RAPID = 'men_rapid',
  WOMEN_RAPID = 'women_rapid',
  JUNIORS_RAPID = 'juniors_rapid',
  GIRLS_RAPID = 'girls_rapid',
  MEN_BLITZ = 'men_blitz',
  WOMEN_BLITZ = 'women_blitz',
  JUNIORS_BLITZ = 'juniors_blitz',
  GIRLS_BLITZ = 'girls_blitz',
}

export enum DataSource {
  CACHE = 'cache',
  SCRAPE = 'scrape',
}

export enum ResourceType {
  LIST = 'list',
  PROFILE = 'profile',
  HISTORY = 'history',
  STATS = 'stats',
}